requirements.txt
blinker==1.3
cffi==1.1.2
cryptography==0.9.3
Cython==0.22.1
Flask==0.10.1
Flask-PyMongo==0.3.1
Flask-Script==2.0.5
idna==2.0
itsdangerous==0.24
Jinja2==2.7.3
jsonschema==2.5.1
Mako==1.0.1
MarkupSafe==0.23
mistune==0.7
pyasn1==0.1.8
pycparser==2.14
PyJWT==1.3.0
pymongo==3.0.3
pytz==2015.4
six==1.9.0
uWSGI==2.0.11
Werkzeug==0.10.4
